requirements.txt
python-dotenv ~= 0.14.0
requests ~= 2.22.0
websocket-client ~= 0.57.0